pyproject.toml
[project]
name = "rpi-digital-signage"
version = "0.1.0"
description = "Raspberry Pi向けデジタルサイネージ - PLC生産データ表示システム"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Moge800", email = "your.email@example.com"}
]
keywords = ["raspberry-pi", "plc", "streamlit", "digital-signage", "production-monitoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Manufacturing",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "plotly>=6.4.0",
    "pydantic-settings>=2.12.0",
    "pymcprotocol>=0.3.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "streamlit-autorefresh>=1.0.1",
    "uvicorn>=0.34.0",
]

[tool.mypy]
ignore_missing_imports = true
cache_dir = ".mypy_cache"
incremental = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
warn_unused_ignores = false

[tool.ruff]
line-length = 88
target-version = "py311"
extend-exclude = [".venv", ".git", "__pycache__", "build", "dist"]

[tool.ruff.lint]
ignore = ["E402"]  # sys.path操作後のインポートを許可

[tool.black]
line-length = 88
target-version = ["py311"]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=6.0.0",
    "black>=24.0.0",
    "ruff>=0.8.0",
    "mypy>=1.19.0",
]
